KConfigDialogManager: Save radio button index in QGroupBox that are composed only by radio buttons
We are suggesting people to port from deprecated KButtonGroup to QGroupBox but the dialog manager does not behave the same. This patch fixes it by assuming that in a groupbox that is composed exclusively by radio buttons and whose config item is an int to save the index of the checked radio button instead of if the group box itself is checked.

KGeography frameworks with KButtonGroup ported to QGroupBox works.

REVIEW: 119698