Switch to breeze or oxygen if they are available and icon theme is either not set or hicolor
Description:
Putting it in this framework is a bit random, but since it's the one that has KStandardAction i thought it was a "good enough place". I'm open to discussion on where to put this.

I am running KGeography in a user which has oxygen icons available but since i'm not running neither the kde QPT nor oxygen style nor anything, while moving from KIcon to QIcon::fromTheme i lost most of my icons (since there is no kiconloader in the middle anymore), and the same for kstandardaction icons.

This patch makes it so that if you are using kstandardactions and your icon theme is hicolor but you have breeze or oxygen installed it changes the theme to one of those.

I am unconvinced if we want this here or if i want this in my (and potentially) every application so when we run in Gnome or Unity, our apps get icons what were getting in kde4 times because they were using KIcon and they would lose now because of the QIcon::fromTheme recommendation.

Obviously the missing icons also get fixed by QT_QPA_PLATFORMTHEME=kde but one can not expect to have this define when on non Plasma desktops and since in the past KIcon gave you oxygen-icons even if you were not on plasma i think it makes sense to do this.

Testing Done:
KGeography menu items have icons again.

Acked by Aleix

REVIEW: 119713