Hide the "Show Menu Bar" action if all the menubars are native
Some applications have a "Show Menu Bar" action that is a bit silly on systems where the menubar is part of the shell (for example Unity 7).

This patch attempts to fix it by iterating all the main windows when they are shown and if all the menubars of all mainwindows are native, then hides the show menu bar action (basically erasing it from existence).

It's not the nicest of the codes and probably has some edge cases but works on the general case so i think it's worth the effort.

Tried konsole, kate and dolphin under Unity 7 on Ubuntu 16.10

konsole and kate work fine (i.e. the action is gone from the menus and all is good)

dolphin is not 100% "perfectly behabed" (i.e. the "control" toolbar item is supposed to not be shown when menubars are shown and in this case it's shown) but it's not a regression and imho it's the dolphin code being a bit weird (i can propose a patch for it if this gets accepted)

REVIEW: 129261