KConfigDialogManager: get change signal from metaObject or special property
Summary:
KConfigDialogManager was possibly designed when there was no way to
access the change signal of properties. So maintaining the signals
and properties to use in separate global user-accessable tables might
have been done as there was no other option.

But this has flaws:
* different libs & plugins might do conflicting global settings
* prevents per-widget-instance settings (for signals)
* requires explicit setup even for signals already set via NOTIFY on the property

With modern Qt metaobject API instead the change signal can be
fetched and used automatically. And if needing a custom signal, e.g. for
classes without a notify signal set on the property, that can be set via a
special new dynamic property "kcfg_propertyNotify", like it is done
to select a custom property for a widget instance.

For backward-compatibility in KF5 those will be only a new options,
the global maps should be removed on next ABI breakage.

Test Plan: Old and new tests do not fail.

Reviewers: #frameworks

Subscribers: elvisangelaccio

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D4130