Relax constraints for processing QGroupBoxes
So far this was only accepting QGroupBoxes containing only auto-exclusive
buttons. Now we also accept QGroupBoxes that contain non-buttons, as long
as the buttons are still all auto-exclusive and direct children of the
group box.

This is closer to what KButtonGroup was doing and enables porting of uses
of those in kdepim-runtime. Group boxes there contain e.g. additional
labels with more explanations.

Review D5344