Deprecate KStandardAction::PasteText and KPasteTextAction
The PasteText standard action creates a KPasteTextAction rather than a
QAction. However KPasteTextAction has not been functional for years
(it's supposed to integrate Klipper in the toolbar, but it doesn't).

We already have KStandardAction::Paste which is widely used,
while according to lxr, there is only one usage of PasteText and 3 of
pasteText(). There is no reason to have two standard actions that do
the same thing, so mark PasteText as deprecated.

Differential Revision: https://phabricator.kde.org/D7294