Avoid overloading KCModule::changed
Summary:
KCModule contains a signal and a protected slot with the same name.
With the new connect syntax this becomes messy, especially as
this method is called from KCMs a lot.

This patch introduces a new name for the slot, with the
intention of dropping the old one in KF6

Reviewers: apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D23512