KCModule: Indicate when a setting has been changed from the default or previous value
Summary:
Extend KConfigDialogManager internals to insert a small
SettingsStatusIndicator widget next to widgets representing a setting
which is differs from default value.

Test Plan:
Tested it with several widgets based KCMs, namely: qtquicksettings,
desktoppath and screenlocker. Those three have different situations
in term of layouts. Worst case scenario is desktoppath which leads
to the right hand side of the fields moving a bit to fit the indicators
but that's the price to pay for the feature I guess while keeping
things readable.

Here is a screenshot how of it looks on the qtquicksettings KCM:

{F8225745}

Note that the indicators are clickable and reset the field to default value.

Reviewers: ngraham, davidedmundson, meven, crossi, bport, #vdg, ndavis, broulik

Subscribers: alexde, ndavis, iasensio, davidre, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27540