Fix the exclusive group box case for default indicators
Summary:
Turns out there was an oversight here, in case of an exclusive group
box, the emitter widget is a button inside the group box which can have
any name (this was thus caught by the assert).

So now we also check if the sender is a button and if that's the case we
verify if its in one of the known exclusive group boxes. If yes we use
the group box as reference for the indicator work.

Reviewers: ngraham, davidedmundson, bport, crossi

Reviewed By: ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29014