kcolorscheme: Properly handle KColorScheme::isColorSetSupported on blank configurations
On a blank configuration, querying isColorSetSupported will always return 'false' when we ask KConfig if a group is present. This causes us to incorrectly return 'false' for all colour sets when the config is blank. By detecting when configuration is blank, we can instead always return 'true as we have a fallback for blank config for all colour schemes.