Introduce KHamburgermenu
This commit introduces the hamburger menu, which is a general menu button,
to frameworks. It substitutes the menu bar meaning that it will act non-
existent when the menu bar is visible. When there is no menu bar or it is
hidden, the KHamburgerMenu will take on all of its responsibilities.

The hamburger menu is supposed to only feature actions which are not
already visible for the user in other parts of the UI. It will try to do
so automatically but methods are available to make it aware of toolbars,
menus, etc.

When both the menu bar and toolbars are hidden, users might find
themselves in a situation they do not know how to recover from. To avoid
this, KHamburgerMenu is supposed to be added to context menus where it
will only be visible, when no other menu is available.

The menu also features a special sub-menu advertising actions which are
only available through the menu bar and would crowd the hamburger menu.

CCBUG: 368421