clarify quickhelp apidocs
- the quickhelp isn't actually used for modern containers such as
systemsettings so to say it is called when docked is wrong, it may be,
or not
- when it is used it'd usually be used for what's this text. that sets
it apart from the other metadata in the desktop file (kinda)

admittedly I am somewhat thinking that this function should actually be
deprecated. it kind of overlaps with `Comment` from the plugin metadata.
it is also fairly rubbish because you'd have to load the plugin to
"what's this" it, but when listing the plugin in e.g. the sidebar you
don't have it necessarily loaded, much less instantiated, making it
worse than `Comment`