Hide KHamburgerMenu in QMenus when a native menu bar is used
When a native menu bar is used, users normally don't need an
additional KHamburgerMenu in any QMenus.

This was already once before the behaviour in this scenario but it
was unnecessarily changed as a side-effect of
901b049ca48e35fe9df07092a8657b1578fad300.

While the m_menuAction can be used as a normal menu by users that
don't mind invoking a QMenu to access any menu actions, its primary
use really is that of a fallback.
Therefore the existence of a native menu bar (no matter what shape
or size it might have) is enough reason for us to hide m_menuAction.

BUG: 439997
FIXED-IN: 5.86