Fix hamburger menu opening as a window on Wayland
Before this commit, if an app window is unfocused and the first
click on the window is the hamburger menu button, the menu would
appear displaced as a separate window on Wayland.

This commit fixes this by setting a sane transientParent for the
menu. This is slightly difficult because there can be any number
of hamburger buttons in any number of different windows so we need
to figure out spontaneously which button was pressed to open the
menu.

BUG: 442375