Remove KSharedConfig from KColorScheme interface and do not use kdeglobals anymore
Instead use a string to specify which colorscheme to use. Also we can use the new
entry points as a switch for the new behavior: The active color scheme is not
read from kdeglobals anymore, the colors file is the single source of truth.
As this relies on Plasma to write a new key pointing to the colorsfile,
its existence is checked and if not there, the old behvaior is used.