[kcolorschememanager] Don't save colors on application start
The KColorSchemeManager ctor loads the saved colorscheme and activates it.

It does that by calling activateScheme(QModelIndex), which calls activateScheme(QString) and saves to the config

The saving is pointless since we are saving the value we just loaded. Instead call activateScheme(QString) directly.

This avoids needless config syncs on startup