Avoid tracking dangling KConfigDialogManager
A manager is created with a parent `new KConfigDialogManager(widget,
config);` this parent comes externally.

KCModules also explicitly deletes managers.

Typically it would have the same parent as the KCModule which is why it
hasn't been a big issue so far.

Having a parent whilst being explicitly managed is an anti-pattern, but
at a minimum we should clear up our list as we go so we don't risk a
double delete.

CCBUG: 456630