KHamburgerMenu: Reset when menu bar contents change
Prior to this change changes to the top-level actions of the menu
bar were not immediately reflected in the contents of the hamburger
menu. This was an oversight based on the idea that menu bars tend
to not change a lot while an application is running.

With this change the menu bar set through setMenuBar() will now be
monitored for changes to its top-level actions and if changes
occur, KHamburgerMenu will be notified that it needs to update its
contents.

Background information: KHamburgerMenu only updates its contents
when that is necessary. This commit fixes one case that was missed
why an update might be needed.