Bind KHamburgerMenu to OpenMainMenu standard shortcut
Since recently KHamburgerMenu provides all the necessary
functionality to open a main menu when activated. It opens the
first menu of the menu bar when that is available, or opens the
hamburger menu itself when the menu bar is not visible/available/
existent.

This commit allows consistent opening of a main menu using the
standard shortcut (currently F10) to to all users of
KHamburgerMenu. Applications can remove this by removing this
shortcut from KHamburgerMenu in the application code.
However, opening a menu when using the standard shortcut is
expected behaviour for accessibility because until now KDE
applications did not provide a consistent way to open a menu,
leaving visually-impaired users unable to e.g. focus the menu bar
and therefore unable to discover/use all the features such
applications had to offer. Check out the commit in KConfig that
introduced the standard shortcut "Open General Menu" for more
details.

Some applications have completely gotten rid of a menu bar and are
only using a hamburger menu. For those it is equally important
that triggering the standard shortcut does consistently open a
menu.