KStandardAction: return name as QString, store as QString-compat raw data
Most API consumers of KStandardAction::name(id) use the returned value
with QString type needing code, and thus convert it to a QString, with
all the overhead.

Storing the data QString-compatible and directly returning a QString
instance spares the extra code needed to convert as well as the runtime
paper cut costs.