Make commandbar a normal widget instead of a window
Commandbar never really needed to be a window on its own because it was
always inside the parent widget. So far it has worked okay but it has
some issues on other platforms like not being able to paste text in
the QLineEdit on macOS.

This change makes the command bar a normal widget that is completely
handled by Qt. This has many advantages:
-> Same behaviour on all platforms
-> Application shortcuts now work from inside the command bar, this
   allows us to jump between various commandbar like widgets in Kate
   directly.
-> faster to open

BUG: 468433