Add KStyleManager::initStyle to enforce proper application style
Add KStyleManager::initStyle(), like https://invent.kde.org/frameworks/kiconthemes/-/merge_requests/136 initTheme() that will apply the widget style configured in the app or kdeglobals with fallback to Breeze.

Will be a nop like initTheme() if we are already running with the proper platform theme to ensure we have good Plasma integration.