Move KMessageBox storage integration plugin from FrameworkIntegration
Reasoning for this is that we want to use KConfig based storage on all
platforms and not only on Plasma to ensure the config are stored in the
same place as the rest of the application.

The default KMessageBox QSettings storage backend should only be used
when KConfig is not available.