KRecentFilesAction: Don't call into QMimeDatabase to get what we already have
There's no point in creating a QMimeType from a name just to then take the name.
All we do, is pass the name onto KActivitiesStats. In theory it might not know
the mime type and give us application/octet-stream but we probably want to know
the original mime type in this case anyway.