KLanguageButton: Propagate tooltip to child button
Calling KLanguageButton::setToolTip() will not change the tooltip
of the actual internal/private QPushButton child widget. This is an
issue because accessibility software might only communicate the
tooltip when it actually belongs to the button getting focus.

This change aims to make the KSwitchLanguageDialog from the KXmlGui
framework accessible by making the tooltip which explains which of
the KLanguageButton is the primary language and which is the
fallback language accessible e.g. to screen readers. Tested with
the Orca screen reader.